requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncode==1.0.0
asyncrcon==1.1.4
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
greenlet==1.1.0
idna==3.2
multidict==5.1.0
psycopg2==2.9.1
python-dotenv==0.18.0
screenutils==0.0.1.6.2
SQLAlchemy==1.3.20
typing-extensions==3.10.0.0
Unidecode==1.2.0
yarl==1.6.3
